pyproject.toml
[project]
name = "hw-web-11"
version = "0.1.0"
description = ""
authors = [
    {name = "Yevhenii Bessonov",email = "norent777@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "python-decouple (>=3.8,<4.0)",
    "fastapi (>=0.115.13,<0.116.0)",
    "alembic (>=1.16.2,<2.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "uvicorn[standart] (>=0.34.3,<0.35.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "pydantic[email] (>=2.11.7,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "libgravatar (>=1.0.4,<2.0.0)",
    "bcrypt (==3.2.2)",
    "fastapi-mail (>=1.5.0,<2.0.0)",
    "fastapi-limiter (>=0.1.6,<0.2.0)",
    "redis (>=6.2.0,<7.0.0)",
    "cloudinary (>=1.44.1,<2.0.0)",
    "sphinx (>=8.2.3,<9.0.0)",
    "sphinx-autodoc-typehints (>=3.2.0,<4.0.0)",
    "sphinx-rtd-theme (>=3.0.2,<4.0.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "pytest-asyncio (>=1.1.0,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
sphinx = "^8.2.3"


[tool.poetry.group.test.dependencies]
httpx = "^0.28.1"
aiosqlite = "^0.21.0"
pytest-asyncio = "^1.1.0"
fakeredis = {extras = ["lua"], version = "^2.30.1"}

[tool.pytest.ini_options]
testpaths = [
    "tests",
]
pythonpath = "."